refactor[FE]: Separate organisms from pages
