+++ frontend/src/components/organisms/Verify2FA.tsx
@@ -37,23 +37,25 @@ export default function Verify2FA({ onSuccess }: Props) {
   };
 
   return (
-    <form onSubmit={handleSubmit} className="w-full">
-      <div className="my-16 text-center">
-        <h2 className="section-title">Two-Factor Authentication</h2>
-        <p>2단계 보안 인증 코드를 입력해 주세요.</p>
-        <div className="my-16 mx-auto">
-          <TextInput
-            id="code"
-            value={code}
-            onChange={handleChange}
-            className="text-3xl text-center w-36 font-mono"
-            placeholder="000000"
-          />
+    <div className="container mx-auto max-w-xl max-h-screen py-24 px-4">
+      <form onSubmit={handleSubmit} className="w-full">
+        <div className="my-16 text-center">
+          <h2 className="section-title">Two-Factor Authentication</h2>
+          <p>2단계 보안 인증 코드를 입력해 주세요.</p>
+          <div className="my-16 mx-auto">
+            <TextInput
+              id="code"
+              value={code}
+              onChange={handleChange}
+              className="text-3xl text-center w-36 font-mono"
+              placeholder="000000"
+            />
+          </div>
+          <Button type="submit" primary fullLength>
+            확인
+          </Button>
         </div>
-        <Button type="submit" primary fullLength>
-          확인
-        </Button>
-      </div>
-    </form>
+      </form>
+    </div>
   );
 }
