+++ frontend/src/pages/MainPage.tsx
@@ -1,42 +1,11 @@
-import { Link } from 'react-router-dom';
-import { useQuery } from '@tanstack/react-query';
-import { getGames } from 'api/api.v1';
-import { ChevronRightIcon } from '@heroicons/react/20/solid';
 import AppTemplate from 'components/templates/AppTemplate';
-import GameButtons from 'components/organisms/GameButtons';
-import ChannelButtons from 'components/organisms/ChannelButtons';
-import GameLives from 'components/organisms/GameLives';
+import Main from 'components/organisms/Main';
 import HeaderGnb from 'components/molecule/HeaderGnb';
 
 export default function MainPage() {
-  const { data: games } = useQuery({
-    queryKey: ['games'],
-    queryFn: getGames,
-  });
-
   return (
     <AppTemplate header={<HeaderGnb />}>
-      <section className="section">
-        <h2 className="section-title">게임</h2>
-        <GameButtons />
-      </section>
-      <section className="section">
-        <div className="flex justify-start items-baseline">
-          <h2 className="section-title">라이브 게임</h2>
-          <span className="inline-block text-sm text-gray-light ml-2">
-            <Link to="/game">더보기</Link>
-            <ChevronRightIcon
-              className="inline-block h-4 w-4"
-              aria-hidden="true"
-            />
-          </span>
-        </div>
-        {games && <GameLives games={games} />}
-      </section>
-      <section className="section">
-        <h2 className="section-title">채널</h2>
-        <ChannelButtons />
-      </section>
+      <Main />
     </AppTemplate>
   );
 }
