+++ frontend/src/pages/Verify2FAPage.tsx
@@ -1,17 +1,11 @@
-import AppTemplateWithoutHeader from 'components/templates/AppTemplateWithoutHeader';
+import AppTemplate from 'components/templates/AppTemplate';
 import Verify2FA from 'components/organisms/Verify2FA';
+import Header from 'components/molecule/Header';
 
 export default function Verify2FAPage() {
   return (
-    <AppTemplateWithoutHeader>
-      <header className="relative mx-auto max-w-xl px-2">
-        <div className="flex h-14 text-lg font-medium text-white vh-center">
-          <a href="/">The Pong</a>
-        </div>
-      </header>
-      <div className="container mx-auto max-w-xl max-h-screen py-24 px-4">
-        <Verify2FA />
-      </div>
-    </AppTemplateWithoutHeader>
+    <AppTemplate header={<Header />}>
+      <Verify2FA />
+    </AppTemplate>
   );
 }
